Remove more or less pointless warnings.

+++ src/libstatgrab/vector.h
@@ -103,7 +103,7 @@ __sg_private void sg_vector_free(struct sg_vector *vector);
 				  (size_t)(new_count), & VECTOR_INIT_INFO(type))
 
 #ifdef STRUCT_SG_VECTOR_ALIGN_OK
-#define VECTOR_SIZE ((size_t)(((struct sg_vector *)0) + 1))
+#define VECTOR_SIZE ((size_t)(((struct sg_vector *)0) + ((size_t)1)))
 #define VECTOR_DATA(vector) \
 	(vector ? (void *)&((vector)[1]) : NULL)
 #define VECTOR_ADDR_ARITH(ptr) \
